backend/requirements.txt
altgraph==0.17.4
annotated-types==0.7.0
anyio==4.7.0
beautifulsoup4==4.12.3
blinker==1.9.0
blis==0.7.11
bs4==0.0.2
cachetools==5.5.0
catalogue==2.0.10
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
cloudpathlib==0.20.0
confection==0.1.5
cymem==2.0.10
dnspython==2.7.0
email_validator==2.2.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl#sha256=86cc141f63942d4b2c5fcee06630fd6f904788d2f0ab005cce45aadb8fb73889
exceptiongroup==1.2.2
fastapi==0.115.6
Flask==3.1.0
flask-cors==5.0.1
google-ai-generativelanguage==0.6.10
google-api-core==2.24.0
google-api-python-client==2.156.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-generativeai==0.8.3
googleapis-common-protos==1.66.0
grpcio==1.69.0
grpcio-status==1.69.0
h11==0.14.0
httplib2==0.22.0
idna==3.10
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.5
langcodes==3.5.0
language_data==1.3.0
logger==1.4
macholib==1.16.3
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
murmurhash==1.0.11
numpy==1.26.4
oauthlib==3.2.2
packaging==24.2
pandas==2.2.3
pluggy==1.5.0
preshed==3.0.9
proto-plus==1.25.0
protobuf==5.29.2
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.4
pydantic_core==2.27.2
Pygments==2.18.0
pyinstaller-hooks-contrib==2024.11
pyparsing==3.2.0
pytest==8.3.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.9.4
rsa==4.9
shellingham==1.5.4
six==1.17.0
smart-open==7.1.0
sniffio==1.3.1
soupsieve==2.6
spacy==3.7.5
spacy-cleaner==3.2.1
spacy-legacy==3.0.12
spacy-loggers==1.0.5
spacy-lookups-data==1.0.5
srsly==2.5.0
starlette==0.41.3
thinc==8.2.5
tomli==2.2.1
tqdm==4.66.6
typer==0.15.1
types-requests==2.32.0.20241016
types-tqdm==4.67.0.20241221
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.34.0
wasabi==1.1.3
weasel==0.4.1
Werkzeug==3.1.3
wrapt==1.17.0
